better implementation for channel preferences

+++ tools/database_tools.py
@@ -1,11 +1,29 @@
 import asyncpg
 import discord
 from discord.ext import commands
+from enum import Enum
 from tools.enum_tools import TableType
 from typing import Union
 
 
-__all__ = ("DatabaseTools",)
+__all__ = (
+    "DatabaseTools",
+    "Actions",
+)
+
+INEPT = object()
+
+
+class Actions(Enum):
+    """
+    Enum for the different actions that can be done with the database
+    switches for ignoring and unignoring channels and commands
+    """
+
+    ignore = 1
+    unignore = 2
+    enable = unignore
+    disable = ignore
 
 
 class DatabaseTools:
@@ -50,7 +68,7 @@ class DatabaseTools:
                 """
         await self.pool.execute(query)
 
-        # Table for storing command preferences 
+        # Table for storing command preferences
         query = """CREATE TABLE IF NOT EXISTS command_preferences(
                     guild_id BIGINT,
                     ignored_command TEXT
@@ -65,8 +83,6 @@ class DatabaseTools:
                     );
                 """
         await self.pool.execute(query)
-
-
 
     async def increment_usage(
         self,
@@ -175,13 +191,15 @@ class DatabaseTools:
                 return False
         return True
 
-    async def is_preferred_command(self, guild_id:int, command_name: str):
+    async def is_preferred_command(self, guild_id: int, command_name: str):
         """
         Returns true if the command is not disabled in the guild
         else returns false
         commands can be re-enabled or disabled using the enable / disable command
         """
-        query = """SELECT ignored_command FROM command_preferences WHERE guild_id = $1;"""
+        query = (
+            """SELECT ignored_command FROM command_preferences WHERE guild_id = $1;"""
+        )
         commands = await self.pool.fetch(query, guild_id)
         if not commands:
             return True
@@ -189,38 +207,71 @@ class DatabaseTools:
             if command.get("ignored_command") == command_name:
                 return False
         return True
-        
 
-    async def ignore_command(self, guild_id: int, command_name: str):
+    async def channel_action(
+        self, ctx, action: Actions, channel: Union[discord.TextChannel, int, str]
+    ):
         """
-        This function is used to ignore a command in the guild
+        This function is used to enable or disable a channel in the guild
         """
-        query = """INSERT INTO command_preferences (guild_id, ignored_command)
-                    VALUES ($1, $2);"""
-        await self.pool.execute(query, guild_id, command_name)
+        # Check list :
+        # 1. If the channel is an int, then assert there's an associated channel (in bot's cache, no fetching)
+        # 2. If the channel is a str, then find the channel then get it's ID
+        # 3. If the channel is a discord.TextChannel, then get it's ID
+        # 4. If the channel is not found then send an appropriate error message
+        # -- At this point we have the channel ID --
+        # 5. If the new action doesn't necessarily do anything to the current state then send the issue message
+        # -- Security --
+        # 6. Check the origin of the channel matches ctx.guild
 
+        # If all of the checks above is passed then we can proceed with the action
 
-    async def unignore_command(self, guild_id: int, command_name: str):
-        """
-        This function is used to unignore a command in the guild
-        """
-        query = """DELETE FROM command_preferences WHERE guild_id = $1 AND ignored_command = $2;"""
-        await self.pool.execute(query, guild_id, command_name)
+        former = (
+            channel[:] if isinstance(channel, discord.discord.TextChannel) else INEPT
+        )
+        assert isinstance(action, Actions)
+        # former is the former value of the channel before it gets changed
+        # Stage 1 -> 4 :
+        if isinstance(channel, int):
+            channel = await self.bot.get_channel(channel)
+            if channel is None:
+                return await ctx.send(f"Cannot find a channel with that ID")
 
+        elif isinstance(channel, str):
+            channel = discord.utils.get(ctx.guild.text_channels, name=channel)
+            if channel is None:
+                return await ctx.send(f"Cannot find a channel named **{former}**")
 
-    async def ignore_channel(self, guild_id: int, channel_id: int):
-        """
-        This function is used to ignore a channel in the guild
-        """
-        query = """INSERT INTO channel_preferences (guild_id, channel_id)
-                    VALUES ($1, $2);"""
-        await self.pool.execute(query, guild_id, channel_id)
+        elif isinstance(channel, discord.discord.TextChannel):
+            pass
 
+        else:
+            raise ValueError(f"{type(channel)} is not a valid type for channel")
 
-    async def unignore_channel(self, guild_id: int, channel_id: int):
-        """
-        This function is used to unignore a channel in the guild
-        """
-        query = """DELETE FROM channel_preferences WHERE guild_id = $1 AND channel_id = $2;"""
-        await self.pool.execute(query, guild_id, channel_id)
-        
+        # Stage 5
+        if await self.is_preferred_channel(
+            ctx.guild.id, channel.id
+        ):  # If true, then the channel is not ignored
+            if action is Actions.unignore:
+                return await ctx.send(
+                    f"{channel.mention} is not ignored to begin with!"
+                )
+        else:  # The channel is ignored
+            if action is Actions.ignore:
+                return await ctx.send(f"{channel.mention} is already ignored")
+
+        # Stage 6
+        if channel not in ctx.guild.text_channels:
+            return await ctx.send(f"**{channel.name}** is not a channel in your guild")
+
+        # -- All checks passed --
+
+        if action is Actions.ignore:
+            query = """INSERT INTO channel_preferences (guild_id, channel_id) VALUES ($1, $2);"""
+            await self.pool.execute(query, ctx.guild.id, channel.id)
+            return await ctx.send(f"{channel.mention} has been ignored")
+
+        elif action is Actions.unignore:
+            query = """DELETE FROM channel_preferences WHERE guild_id = $1 AND channel_id = $2;"""
+            await self.pool.execute(query, ctx.guild.id, channel.id)
+            return await ctx.send(f"{channel.mention} has been unignored")
